Add new file for handling stripe payments. Navigates to stripes checkout webpage instead of embedded. - Fix backend route/controller to handle the new changes and send the right data.
Introduce a new .tsx file that is taken from my other projects as i was
hving lots of issues with the embedded version and wasn't worth the time
messing with the frontend as the backend data was correctly being passed
through. Use Stripes dedicated checkout page instead, Now Payments are
working and we can process orders now.