add dotenv to bullmq and into the relevant files. add our stripe product creation into our worker, now our worker handles all of the stripe logic correctly and away from our product services controllers. add StripeProduct wrapper functions. -> Product creation route tested in requestly.
Product creation flow works perfectly in our controller, handle our null
checks from the incoming data. add the product to the database and then
we add our data to a new job for our worker to process and then add that
product to our stripe catalog ready for later validation when processing
payments.