add our mongoose connection check in our createOrder function.
This is the wrong place to put it, need to refactor it tomorrow and put
it inside our worker as we get errors due to the string being undefined
momentarily because our worker tries to process everyhthing before our
MONGO_URL is defined and tehn finally processes, so we need to add those
checks into our worker so it dosen't process anything until connected.
Check BullMq docs.